[REFACTOR] sendImagePostRequest 모듈화
/common/api.js로 이동
